requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.2
whitenoise==6.2.0
